test/pylib: shorten UNIX-domain sock paths
The "pylib" library we use in the topology test suite uses a UNIX-domain
socket to communicate between the cluster manager and the individual
tests. The socket is located in the test directory (where both server
and client can easily find it) but there is a problem: In Linux the
length of the path used as a UNIX-domain socket address is limited to
just a little over 100 bytes. In Jenkins run, the test directory names
are very long, and we sometimes go over this path-size limit and the
result is that test.py fails creating this socket.

In this path we add a simple "socket-path shortening" class,
which temporarily gives us a shorter name for the long path, by
opening the directory containing it and then referring to it using
"/proc/self/fd/NUM". We then use this socket-path shortener in the
two places in the code that open the UNIX-domain socket (the server
and the client).

Tested by cloning Scylla in a very long directory name. A test like
`./test.py --mode=dev test_concurrent_schema` fails before this patch,
and passes with it.

Fixes #12622

Signed-off-by: Nadav Har'El <nyh@scylladb.com>