cql3: expr: validate arguments to the token() function
The token() function is very strict in what is acceptable
as its arguments.
The only valid way to call token() is by passing all
partition key columns as arguments, in exactly
the same order as in the table definition.

So for a table like:
CREATE TABLE (p1 int, p2 int, c int, r int, PRIMARY KEY ((p1, p2), c));

The only valid way to call token() is by typing: token(p1, p2)

The general expectation is that prepare_expression() should handle
most type checking and validation, so it's reasonable to expect
that it would also validate arguments for the token() function call.

This wasn't the case up until now. Because of the way the code evolved
it was never needed to perform this vaidation here, so it was never
implemented. As a part of moving everything to prepare_expression()
and evaluate() it needs to be implemented sooner or later.

Currently token argument validation takes place in statement_restrictions,
but it would be better to have it work on generic expressions and not only on
the WHERE clause. If this were the case and selectable was based on prepared
expressions we wouldn't have bugs like:
https://github.com/scylladb/scylladb/issues/10448

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>