cql3: expr: add receiver validation when preparing token()
prepare_expression() has an optional argument called receiver.
This argument specifies what's the type of the value that the prepard expression will be assigned to.
Generally prepare_expression should check that the type is correct,
e.g we're not trying to assign an int to a text value.
The code evolved in a way where some values don't have these checks because
everything works without them for now.

I'm adding the check because it's the right thing to do
if we want prepare_expression to achieve everything
it was meant to. It allows me to write the tests
that test the generally expected behavior instead
of creating some specific cases to match the existing implementation.

This doesn't fix any existing bugs, but in the future
it will be part of the validation code so it's important to have.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>