dist/debian: drop unused Makefile variable
`job` was introduced back in 782ebcece4f860238abc4dbc5dc4f81fdd1f02f9,
so we could consume the option specified in DEB_BUILD_OPTIONS
environmental variable. but now that we always repackage
the artifacts prebuilt in the relocatable package. we don't build
them anymore when packaging debian packages. see
9388f3d626ce9f366f5fa16c2a683c419e427cc2 . and `job` is not
passed to `ninja` anymore.

so, in this change, `job` is removed from debian/rules as well, as
it is not used.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>