compress: Add info + helpers for read/write compressed packet uncompressed size to data
This is OCD level stuff, but because we for historical reasons (i.e. cassandra sstable compatibility)
write uncompressed size in some compressors (hello lz4), I cannot in good concience allow a user (me)
to duplicate this info for other. So we keep track per compressor type etc whether and how to add
this to a packet.

This is a helper to be able to write "other" usage of compressor with minimal overhead.